% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_parameters.R
\name{Augment_Missing}
\alias{Augment_Missing}
\title{Augment missing response with observed information}
\usage{
Augment_Missing(data)
}
\arguments{
\item{data}{Dataset contain missing response indicated as'NA', including response variable as 'y' and dose variable as 'dose'.}
}
\value{
A complete dataset with augmentation.
}
\description{
Augment missing response with 0 and 1, and remain all other variables the same.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
