% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_hess.R
\name{Comp_I}
\alias{Comp_I}
\title{Compute analytical form of expected information matrix of Binary Emax model}
\usage{
Comp_I(data, weight, theta)
}
\arguments{
\item{data}{A complete dataset without missingness, including response variable as 'y' and dose variable as 'dose'.}

\item{weight}{A vector of working weight from the EM iterations.}

\item{theta}{Parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).}
}
\value{
A matrix of expected information
}
\description{
Compute expected information matrix of Binary Emax model
}
\details{
Compute expected information matrix of Binary Emax model which will be used in finding derivative of Jeffery's prior.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
