% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_parameters.R
\name{comp_Q_firth}
\alias{comp_Q_firth}
\title{Maximization function estimation of bias reduced EM algorithm with defined weight.}
\usage{
comp_Q_firth(data, theta, alpha, weight, fit_mis)
}
\arguments{
\item{data}{A complete dataset without missingness, including response variable as 'y' and dose variable as 'dose'.}

\item{theta}{Parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).}

\item{alpha}{Parameters of logisitic missing indicator model.}

\item{weight}{A vector of working weight from the EM iterations.}

\item{fit_mis}{an object of class "formula": a symbolic description of the model to be fitted.}
}
\value{
A value of function estimation
}
\description{
Estimate Maximization function with given parameters, weight, and data for the bias reduced EM Emax model.
}
\details{
Calculating the maximization function of bias reduced EM algorithm for each iteration.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
