% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_parameters.R
\name{comp_theta}
\alias{comp_theta}
\title{Estimation of emax parameters in EM algorithm iteration.}
\usage{
comp_theta(data, weight, theta)
}
\arguments{
\item{data}{A complete dataset without missingness, including response variable as 'y' and dose variable as 'dose'.}

\item{weight}{A vector of working weight from the EM iterations.}

\item{theta}{Initial value of parameters for optimization. The order of the variables is (E0,log(ED50),Emax).}
}
\value{
A vector of parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).
}
\description{
Calls Newton-Raphson optimizer MaxNR, for Emax model.
}
\details{
Fits the Emax model with defined working weights using MaxNR.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
