% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_vcov.R
\name{comp_vcov}
\alias{comp_vcov}
\title{Calculate the variance covariance matrix of estimated parameters by EmaxEM}
\usage{
comp_vcov(em.emax.fit)
}
\arguments{
\item{em.emax.fit}{an object for result from EmaxEM}
}
\value{
A list of two variance covariance matrices, one for Emax model parameter, one for logistic missing model parameter.
}
\description{
This provides the estimated VCOV matrix of parameters using IL method by EmaxEM.
}
\details{
Internal function for variance covariance estimation.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MASS]{ginv}}
 \code{\link[boot]{inv.logit}}
}
