% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emaxem_main.R
\name{fitEmaxEM_firth}
\alias{fitEmaxEM_firth}
\title{Fitting bias reduced IL method with Emax model and binary response missing data.}
\usage{
fitEmaxEM_firth(
  data = NULL,
  theta_0 = NULL,
  alpha_0 = NULL,
  mis_form = as.formula(mis ~ y + dose + x1 + x2)
)
}
\arguments{
\item{data}{Dataset contain missing response indicated as'NA', including response variable as 'y' and dose variable as 'dose'.}

\item{theta_0}{Initial value of Emax model parameters for optimization, Default: NULL}

\item{alpha_0}{Initial value of logistic missingness model parameters for optimization., Default: NULL}

\item{mis_form}{an object of class "formula": a symbolic description of the model to be fitted, Default: as.formula(mis ~ y + dose + x1 + x2)}
}
\value{
list of fitted values:
\item{theta}{the final fitted parameters of Emax model}
\item{alpha}{the final fitted parameters of logistic missing model}
\item{weight}{the final fitted weight for each observation in EM}
\item{Q}{the value of Q function for maximizationfor each iteration of EM}
\item{K}{the total number of iterations of EM to converge}
\item{vcov_theta}{the estimated variance covariance matrix of theta}
\item{vcov_alpha}{the estimated variance covariance matrix of alpha}
}
\description{
This provides the estimates using FIL method.
}
\examples{
\dontrun{
if(interactive()){
 theta_true=matrix(c(qlogis(0.1),qlogis(0.8)-qlogis(0.1),log(7.5)),1,3)
 colnames(theta_true)<- c('e_0','emax','led_50')
 theta_true <- as.data.frame(theta_true)
 dose_set <- c(0,7.5,22.5,75,225)
 n=355
 alpha_true = c(0.5,1,-0.5,0,0) #mis 15 typical
 data <-sim_data(theta_true,n,dose_set,alpha_true)
 res <- fitEmaxEM_firth(data=data$data,mis_form=as.formula(mis~y+dose) )
 }
}
}
\seealso{
\code{\link[clinDR]{fitEmax}}
}
