% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_parameters.R
\name{log_Emax_i}
\alias{log_Emax_i}
\title{Log likelihood estimation of binary Emax model}
\usage{
log_Emax_i(data, theta)
}
\arguments{
\item{data}{A complete dataset without missingness, including response variable as 'y' and dose variable as 'dose'.}

\item{theta}{Parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).}
}
\value{
A vector of log likelihood of each observation.
}
\description{
Estimate Log likelihood of given parameters with data for binary Emax model.
}
\details{
Internal function of calculating the maximization function of EM algorithm.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
