% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_parameters.R
\name{log_missing_i}
\alias{log_missing_i}
\title{Log likelihood estimation of logisitic missing indicator model}
\usage{
log_missing_i(data, alpha, mis_form)
}
\arguments{
\item{data}{A complete dataset without missingness, including response variable as 'y' and dose variable as 'dose'.}

\item{alpha}{Parameters of logisitic missing indicator model.}

\item{mis_form}{an object of class "formula": a symbolic description of the model to be fitted.}
}
\value{
A vector of log likelihood of each observation.
}
\description{
Estimate Log likelihood of given parameters with data for logisitic missing indicator model.
}
\details{
Internal function of calculating the maximization function of EM algorithm.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
