% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate dataset for testing Emaxem and Emaxem_firth}
\usage{
sim_data(theta, n, dose_set, alpha)
}
\arguments{
\item{theta}{True parameters of Emax model. The order of the variables is (E0,log(ED50),Emax).}

\item{n}{number of observations.}

\item{dose_set}{A vector indicate the dose set for the dose-response relationship.}

\item{alpha}{True parameters of logisitic missing model.}
}
\value{
A list of two datasets. One is with missingness on repsonse, and the other is the full complete data.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 theta_true=matrix(c(qlogis(0.1),qlogis(0.8)-qlogis(0.1),log(7.5)),1,3)
colnames(theta_true)<- c('e_0','emax','led_50')
theta_true <- as.data.frame(theta_true)
dose_set <- c(0,7.5,22.5,75,225)
n=355
alpha_true = c(0.5,1,-0.5,0,0) #mis 15 typical
data <-sim_data(theta_true,n,dose_set,alpha_true)
 }
}
}
